+++ crates/uselesskey-core-x509-negative/README.md
@@ -1,11 +1,12 @@
 # uselesskey-core-x509-negative
 
-Policy-only X.509 negative-fixture helpers.
+Policy-only X.509 certificate negative-fixture helpers.
 
 ## Purpose
 
-- Keep all X.509 negative fixture policies in one crate (`X509Negative`, `ChainNegative`).
+- Keep certificate-level X.509 negative fixture policies in one crate (`X509Negative`).
 - Provide deterministic, stable helpers for negative-variant mutation and metadata.
+- Re-export `ChainNegative` from `uselesskey-core-x509-chain-negative` for compatibility.
 - Stay intentionally free of encoding/parsing/serialization logic.
 
 ## Responsibilities
@@ -15,9 +16,3 @@ Policy-only X.509 negative-fixture helpers.
   - not-yet-valid cert fixtures
   - wrong key usage policy variants
   - CA-policy mismatch variants
-- `ChainNegative`:
-  - chain-level hostname mismatch
-  - unknown root CA identity
-  - expired leaf/intermediate cert variants
-  - revoked leaf variant metadata hooks
-
